+++ pymc3/tests/test_distributions_random.py
@@ -375,6 +375,11 @@ class TestLaplace(BaseTestCases.BaseTestCase):
     params = {"mu": 1.0, "b": 1.0}
 
 
+class TestAsymmetricLaplace(BaseTestCases.BaseTestCase):
+    distribution = pm.AsymmetricLaplace
+    params = {"kappa": 1.0, "b": 1.0, "mu": 0.0}
+
+
 class TestLognormal(BaseTestCases.BaseTestCase):
     distribution = pm.Lognormal
     params = {"mu": 1.0, "tau": 1.0}
@@ -625,6 +630,17 @@ class TestScalarParameterSamples(SeededTest):
             return st.laplace.rvs(mu, b, size=size)
 
         pymc3_random(pm.Laplace, {"mu": R, "b": Rplus}, ref_rand=ref_rand)
+
+    def test_laplace_asymmetric(self):
+        def ref_rand(size, kappa, b, mu):
+            u = np.random.uniform(size=size)
+            switch = kappa ** 2 / (1 + kappa ** 2)
+            non_positive_x = mu + kappa * np.log(u * (1 / switch)) / b
+            positive_x = mu - np.log((1 - u) * (1 + kappa ** 2)) / (kappa * b)
+            draws = non_positive_x * (u <= switch) + positive_x * (u > switch)
+            return draws
+
+        pymc3_random(pm.AsymmetricLaplace, {"b": Rplus, "kappa": Rplus, "mu": R}, ref_rand=ref_rand)
 
     def test_lognormal(self):
         def ref_rand(size, mu, tau):